Move to physics atmosphere folder and rename to rayleigh scattering
