+++ src/ts/utils/physics/atmosphere/rayleighScattering.ts
@@ -15,7 +15,7 @@
 //  You should have received a copy of the GNU Affero General Public License
 //  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import { DeepReadonly } from "./types";
+import { DeepReadonly } from "../../types";
 
 /**
  * Boltzmann constant, J K⁻¹  (CODATA 2019)
